\name{crocs}
\alias{crocs}
\docType{data}
\title{
Crocodile morphometrics
}
\description{
A linear morphometric dataset featuring 23 cranial measurements from 223 specimens representing 21 crocodilian species.
}
\usage{data(crocs)}
\format{
 A n X m dataframe, where n is the number of specimens and m is the number of variables.
  }
\source{
http://datadryad.org/resource/doi:10.5061/dryad.m01st7p0
}
\references{
Brown, C., Arbour, J. and Jackson, D. 2012. Testing of the Effect of Missing Data Estimation and Distribution in Morphometric Multivariate Data Analyses. \emph{Systematic Biology} 61(6):941-954.
}
\seealso{
\code{\link{obliterator}}, \code{\link{byclade}},\code{\link{missing.data}},\code{\link{crocs.landmarks}}
}

