\name{dacrya}
\alias{dacrya}
\docType{data}
\title{
Landmark data from \emph{Guianacara dacrya}
}
\description{
Sixteen landmarks taken from the lateral profile of 73 specimens from the Essequibo and rio Branco drainages, used in the description of \emph{Guianacara dacrya}
}
\usage{data(dacrya)}
\format{
A n * l X 2 matrix, with n being the number of specimens and l being the number of landmarks
  }
\source{
Arbour, J. and Lopez-Fernandez, H. 2011. \emph{Guiancara dacrya}, a new species from the rio Branco and Essequibo River drainages
of the Guiana Shield (Perciformes: Cichlidae). \emph{Neotropical Ichthyology} 9:87-96.
}
\seealso{
\code{\link{align.missing}}, \code{\link{MissingGeoMorph}}
}