\name{LOST-package}
\alias{LOST-package}
\alias{LOST}
\docType{package}
\title{
Missing morphometric data simulation and estimation
}
\description{
LOST includes functions for simulating missing morphometric data randomly, with taxonomic bias and with anatomical bias as described by Brown et al. \emph{In Press}. This package also includes functions for estimating missing morphometric data based on regression analysis and a function for checking the percentage of missing data in a matrix.
}
\details{
\tabular{ll}{
Package: \tab LOST\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-09-26\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
J. Arbour and C. Brown

Maintainer: jessica.arbour@utoronto.ca
}
\references{
Arbour, J. and Brown, C. \emph{In Press}. Incomplete specimens in Geometric Morphometric Analyses. \emph{Methods in Ecology and Evolution}

Brown, C., Arbour, J. and Jackson, D. 2012. Testing of the Effect of Missing Data Estimation and Distribution in Morphometric Multivariate Data Analyses. \emph{Systematic Biology} 61(6):941-954.
}

