\name{remove.dat}
\alias{remove.dat}
\title{
Internal function for by.clade
}
\description{
Internal function
}
\usage{
remove.dat(specimen, removes)
}
\arguments{
  \item{specimen}{
specimen to have data removed
}
  \item{removes}{
number of measurements to be removed
}
}

\seealso{
\code{\link{byclade}}
}
