\name{remove.points}
\alias{remove.points}
\title{
Internal function for obliterator
}
\description{
Internal function
}
\usage{
remove.points(specimen, r, distances, expo)
}
\arguments{
  \item{specimen}{
specimen to have data points removed
}
  \item{r}{
number of data points to remove
}
  \item{distances}{
reference specimen landmarks
}
  \item{expo}{
an optional exponent for increasing or decreasing the intensity of anatomical bias
}
}
\seealso{
\code{\link{obliterator}}
}
